Update some rail attribute names, remove others (#114)
https://linear.app/terminal49/issue/DATA-3123/review-and-update-rail-api-documentation

We're renaming a bunch of properties in the API, based on https://github.com/Terminal49/tnt-api/pull/7480/files

@mattyturner also requested the following as part of the ticket:

Remove fees / holds / yard location

We are still sending empty holds and fees, but I think we just want to remove the documentation? I've gone ahead and removed all references. Maybe we split up these two things if we're not sure how far we want to go with removing holds and fees.

-----

* Remove holds, fees, and yard_location attributes for pickup_facility

* inland_destination -> ind, pickup_facility -> ind_facility

* Remove all references to holds and fees